+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 
 # Unreleased
  - Add support for transient services
+ - Rename config file to .candle.json
+ - add-service: Autocreate the config file if missing
 
 # 0.9.0
  - Show observed logs when wait-for-log fails.